feat(s2-x-1): scaffold schema-redesign-impl

+++ app/schemas/account.py
@@ -4,7 +4,6 @@ from __future__ import annotations
 
 import uuid
 from datetime import datetime
-from decimal import Decimal
 
 from pydantic import BaseModel
 
@@ -16,8 +15,10 @@ from app.models.account import AccountType
 
 
 class AccountCreate(BaseModel):
+    # ✍️ code: str  — Chart of Accounts code e.g. "1100", "2000"
     name: str
     account_type: AccountType
+    # ✍️ currency: str  — ISO 4217 code e.g. "EUR"
 
 
 # ---------------------------------------------------------------------------
@@ -27,9 +28,13 @@ class AccountCreate(BaseModel):
 
 class AccountRead(BaseModel):
     id: uuid.UUID
+    # ✍️ code: str
     name: str
     account_type: AccountType
+    # ✍️ currency: str
+    # ✍️ is_active: bool
     created_at: datetime
+    # ✍️ updated_at: datetime
 
     model_config = {"from_attributes": True}
 
@@ -40,5 +45,5 @@ class AccountRead(BaseModel):
 
 
 class BalanceResponse(BaseModel):
-    balance: Decimal
+    balance: int  # BIGINT minor units e.g. 1000 = €10.00; changed from Decimal
     as_of: datetime
